duplicate users Fixes #51

+++ app/routes.py
@@ -64,7 +64,8 @@ def register():
 
     if form.validate_on_submit():
         hashed_password = bcrypt.generate_password_hash(form.password.data)
-        new_user = User(username=form.username.data, password=hashed_password)
+        username = form.username.data.lower()
+        new_user = User(username=username, password=hashed_password)
         db.session.add(new_user)
         db.session.commit()
         return redirect(url_for('login'))
@@ -95,16 +96,16 @@ def delete_file(filename):
 @app.route('/deleteUser/<username>')
 def delete_user(username):
     userr = User.query.filter_by(username=username).first()
-    print(userr)
-    if userr.uploads:
-        files = userr.uploads
-        print(11)
-        for file in files:
-            db.session.delete(file)
-            db.session.commit()
-    db.session.delete(userr)
-    db.session.commit()
-    flash('User deleted!')
+    if userr:
+        if userr.uploads:
+            files = userr.uploads
+            print(11)
+            for file in files:
+                db.session.delete(file)
+                db.session.commit()
+        db.session.delete(userr)
+        db.session.commit()
+        flash('User deleted!')
     if current_user == 'kamalkoranga13+9gse6':
         return redirect(url_for('admin'))
     else:
